+++ src/cosmology/api/__init__.py
@@ -11,7 +11,7 @@ from cosmology.api._components import (
     HasPhotonComponent,
 )
 from cosmology.api._constants import CosmologyConstantsAPINamespace
-from cosmology.api._core import CosmologyAPI
+from cosmology.api._core import Cosmology
 from cosmology.api._extras import HasHubbleParameter, HasTcmb
 from cosmology.api._namespace import CosmologyAPINamespace
 from cosmology.api._standard import StandardCosmology
@@ -21,7 +21,7 @@ from cosmology.api.compat import (
 )
 
 __all__ = [
-    "CosmologyAPI",
+    "Cosmology",
     "FriedmannLemaitreRobertsonWalker",
     "StandardCosmology",
     # components