+++ CHANGES.rst
@@ -4,7 +4,7 @@ Changelog (``cosmology.api``)
 0.1.0 (unreleased)
 ------------------
 
-- CosmologyAPI is the Protocol defining conformant cosmology classes.
+- Cosmology is the Protocol defining conformant cosmology classes.
   (#1) [nstarman]
 
 -  CosmologyAPINamespace is the Protocol defining the main namespace of the